# Inline and optimize `push!()` when one-shot histogramming