--- a/docs/src/notebooks/makie_plotting.jl
+++ b/docs/src/notebooks/makie_plotting.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.5
+# v0.19.32
 
 using Markdown
 using InteractiveUtils
@@ -175,7 +175,7 @@ with_theme(ATLASTHEME) do
 end
 
 # ╔═╡ Cell order:
-# ╟─4cf8f502-eba2-11ec-275f-3f1858e4c2e6
+# ╠═4cf8f502-eba2-11ec-275f-3f1858e4c2e6
 # ╠═82a6fb18-e9c9-450d-9c93-e05bd6cf5859
 # ╠═180e65b9-13db-4b0c-8b16-369b6978cec0
 # ╠═8491f5c3-93b6-4670-8f6e-0d08d7afbf75