--- a/src/hist1d.jl
+++ b/src/hist1d.jl
@@ -140,7 +140,7 @@ Base.broadcastable(h::Hist1D) = Ref(h)
     Hist1D(elT::Type{T}=Float64; bins, overflow) where {T}
 
 Initialize an empty histogram with bin content typed as `T` and bin edges.
-To be used with [`push!`](@ref). Default overflow behavior (`false`)
+To be used with [`Base.push!`](@ref). Default overflow behavior (`false`)
 will exclude values that are outside of `binedges`.
 """
 function Hist1D(elT::Type{T}=Float64; bins, overflow=_default_overflow) where {T}