--- a/src/hist3d.jl
+++ b/src/hist3d.jl
@@ -107,7 +107,7 @@ Base.broadcastable(h::Hist3D) = Ref(h)
     Hist3D(elT::Type{T}=Float64; binedges, overflow) where {T}
 
 Initialize an empty histogram with bin content typed as `T` and bin edges.
-To be used with [`Base.push!`](@ref). Default overflow behavior (`false`)
+To be used with `push!` or [`atomic_push!`](@ref). Default overflow behavior (`false`)
 will exclude values that are outside of `binedges`.
 """
 function Hist3D(elT::Type{T}=Float64; bins, overflow=_default_overflow) where {T}