## Before
```julia
julia> @benchmark Hist1D(x, range(-1,2;length=31)) setup=x=rand(10000000)
BenchmarkTools.Trial: 85 samples with 1 evaluation.
 Range (min … max):  42.596 ms … 181.668 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     47.884 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   48.218 ms ±  15.178 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▅▁
  ███▆▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▃▇▆▄▃▃▁▁▄▁▃▁▁▁▁▁▁▁▁▅▆▆▆▃▁▁▁▃▁▃▄ ▁
  42.6 ms         Histogram: frequency by time         53.7 ms <

 Memory estimate: 1.00 KiB, allocs estimate: 6.
```


## This PR
```julia
julia> @benchmark Hist1D(x, range(-1,2;length=31)) setup=x=rand(10000000)
BenchmarkTools.Trial: 207 samples with 1 evaluation.
 Range (min … max):  12.546 ms …  13.519 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     12.810 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.819 ms ± 164.795 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ▂▁█▁ ▃ ▃    ▁ ▂▃▅    ▅▁▂  ▁
  ▄▁▄▃▄████▆█▆█▇▇▄▇█▄███▅▅▇▆███▄██▃▅▄▄▄▄▄▄▁▄▁▃▁▁▃▃▁▃▁▁▁▁▁▃▃▁▃▃ ▄
  12.5 ms         Histogram: frequency by time         13.3 ms <

 Memory estimate: 1.00 KiB, allocs estimate: 6.
```

## Compare to [fast-histogram](https://github.com/astrofrog/fast-histogram/) (C backend) and [boost-histogram](https://github.com/scikit-hep/hist) (C++ backend)

```python
In [1]: from hist import Hist

In [7]: import numpy as np

In [8]: x = np.random.random(10_000_000)

In [13]: from fast_histogram import histogram1d

In [12]: %%timeit
    ...: h = Hist.new.Reg(30, -1, 2).Int64()
    ...: h.fill(x)
    ...:
    ...:
21.1 ms ± 71.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [14]: %timeit _ = histogram1d(x, range=[-1, 2], bins=30)
13.9 ms ± 53.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```