@Moelf 2023-11-06
@bojohnson5 are you interested in performance aspects of this package?

@rayegun I feel like somehow you might care. And actually you might be able to tell me how to make this more maintainable